Script for batch updates to CSPROJ files (#1)
Let's put together a script that updates csproj files in bulk.  This script will try to make a change to a csproj file, and if successful, it will test the project to see if the change can still be built.  If everything passes, the script saves all the necessary changes.